Moved shared errors into src/shared/errors; refactored cleanUpFilter
