+++ src/client/driver/command-executors/browser-manipulation/ensure-crop-options.js
@@ -1,7 +1,7 @@
 import { styleUtils } from '../../deps/testcafe-core';
 import { getOffsetOptions } from '../../deps/testcafe-automation';
 import limitNumber from '../../../../utils/limit-number';
-import { ActionInvalidScrollTargetError, InvalidElementScreenshotDimensionsError } from '../../../../errors/test-run/shared';
+import { ActionInvalidScrollTargetError, InvalidElementScreenshotDimensionsError } from '../../../../shared/errors';
 
 
 function determineDimensionBounds (bounds, maximum) {
