+++ src/client/driver/driver-link/window/parent.js
@@ -1,6 +1,6 @@
 import { SetAsMasterMessage } from '../messages';
 import sendMessageToDriver from '../send-message-to-driver';
-import { CannotSwitchToWindowError } from '../../../../errors/test-run/shared';
+import { CannotSwitchToWindowError } from '../../../../shared/errors';
 import { WAIT_FOR_WINDOW_DRIVER_RESPONSE_TIMEOUT } from '../timeouts';
 
 export default class ParentWindowDriverLink {
