+++ src/shared/errors/child-window-error-factory.js
@@ -1,10 +1,10 @@
-import { TEST_RUN_ERRORS } from '../types';
+import { TEST_RUN_ERRORS } from '../../errors/types';
 
 import {
     CannotCloseWindowWithChildrenError,
     SwitchToWindowPredicateError,
     WindowNotFoundError
-} from './shared';
+} from './';
 
 export default class ChildWindowValidationErrorFactory {
     static createError ({ errCode, errMsg }) {
