+++ src/shared/errors/index.js
@@ -2,7 +2,7 @@
 // WARNING: this file is used by both the client and the server.
 // Do not use any browser or node-specific API!
 // -------------------------------------------------------------
-import { TEST_RUN_ERRORS } from '../types';
+import { TEST_RUN_ERRORS } from '../../errors/types';
 
 // Base
 //--------------------------------------------------------------------
