+++ src/test-run/commands/options.js
@@ -15,7 +15,7 @@ import {
     ActionPositiveIntegerOptionError,
     ActionBooleanOptionError,
     ActionSpeedOptionError
-} from '../../errors/test-run/shared';
+} from '../../shared/errors';
 
 export const integerOption         = createIntegerValidator(ActionIntegerOptionError);
 export const positiveIntegerOption = createPositiveIntegerValidator(ActionPositiveIntegerOptionError);
