+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/array
@@ -20,7 +20,6 @@ AssertionError: expected [ 1, 2, 3, [ 4, 5 ], 6 ] to deeply equal [ 1, 4, 2,
  ]
 
 
-
 Browser: Chrome 15.0.874.120 / macOS 10.15
 Screenshot: /unix/path/with/<tag>
