+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/boolean
@@ -3,9 +3,7 @@ AssertionError: expected true to deeply equal false
 + expected - actual
 
 -true
-
 +false
-
 
 
 Browser: Chrome 15.0.874.120 / macOS 10.15
