+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/empty-representation
@@ -3,9 +3,7 @@ AssertionError: expected [] to deeply equal ''
 + expected - actual
 
 -[]
-
 +''
-
 
 
 Browser: Chrome 15.0.874.120 / macOS 10.15
