+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/function
@@ -8,7 +8,6 @@ AssertionError: expected [Function] to deeply equal [Function]
          }
 
 
-
 Browser: Chrome 15.0.874.120 / macOS 10.15
 Screenshot: /unix/path/with/<tag>
