+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/string
@@ -3,9 +3,7 @@ AssertionError: expected 'hi' to deeply equal 'hey'
 + expected - actual
 
 -'hi'
-
 +'hey'
-
 
 
 Browser: Chrome 15.0.874.120 / macOS 10.15
